added backend lost in action

+++ src/main/java/com/beerandbeverages/backend/quiz/dto/QuizAnswerDto.java
@@ -1,7 +1,7 @@
 package com.beerandbeverages.backend.quiz.dto;
 
 public record QuizAnswerDto(
-        Long answerId,
+        Integer answerId,
         String key,
         String text
 ) {}
