+++ src/main/java/com/beerandbeverages/backend/quiz/dto/QuizSubmitRequest.java
@@ -3,7 +3,8 @@ package com.beerandbeverages.backend.quiz.dto;
 import java.util.Map;
 
 public record QuizSubmitRequest(
-        Map<Long, Long> answers, // questionId -> answerId
+        Map<Integer, Integer> answers,
         String gender
 ) {}
+
