+++ src/main/java/com/beerandbeverages/backend/quiz/repository/QuizQuestionRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
 public interface QuizQuestionRepository
-        extends JpaRepository<QuizQuestion, Long> {
+        extends JpaRepository<QuizQuestion, Integer> {
 
     List<QuizQuestion> findAllByOrderByPositionAsc();
 }
