+++ src/main/java/com/beerandbeverages/backend/quiz/repository/RaceRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RaceRepository extends JpaRepository<Race, Long> {
+public interface RaceRepository extends JpaRepository<Race, Integer> {
 }
